+++ src/HaikuApi/Services/IUserService.cs
@@ -5,8 +5,8 @@ namespace HaikuApi.Services;
 public interface IUserService
 {
     Task<IEnumerable<User>> GetAllUsersAsync();
-    Task<User?> GetUserByIdAsync(int id);
+    Task<User?> GetUserByIdAsync(short id);
     Task<User> CreateUserAsync(User user);
-    Task<User?> UpdateUserAsync(int id, User user);
-    Task<bool> DeleteUserAsync(int id);
+    Task<User?> UpdateUserAsync(short id, User user);
+    Task<bool> DeleteUserAsync(short id);
 }
